changed random task results sorting key from int to long to get around value size limit issues for over 10 iterations (testing working with 100000 iterations)

+++ ActivitySimulator/Program.cs
@@ -94,7 +94,7 @@ namespace ActivitySimulator
         {
             var results = AssembleActivityDurationOperationsRandom(activities, iterations);
 
-            results = results.OrderBy(operationActivities => int.Parse(string.Join(string.Empty, operationActivities.Select(operationActivity => operationActivity[1])))).ToList();
+            results = results.OrderBy(operationActivities => long.Parse(string.Join(string.Empty, operationActivities.Select(operationActivity => operationActivity[1])))).ToList();
 
             CalculateExpectedDuration(activities, results);
         }
